added data.js and fixed route comments
