+++ front-end/src/components/OneReview.js
@@ -2,26 +2,14 @@
 
 import DeleteButton from "./DeleteButton"
 import UpdateButton from "./UpdateButton"
+import hotelData from "../data"
 
 
 
 
 
 const OneReview =(props)=>{
-    const hotelData= [
-        {value : 0 , label:'Grand Hotel'},
-        {value : 1 , label:'Ocean View Resort'},
-        {value : 2 , label:'Mountain Lodge'},
-        {value : 3 , label:'Lakeside Inn'},
-        {value : 4 , label:'Sunset Resort'},
-        {value : 5 , label:'Ski Chalet'},
-        {value : 6 , label:'Beachfront Paradise'},
-        {value : 7 , label:'City Center Hotel'},
-        {value : 8 , label:'Riverside Retreat'},
-        {value : 9 , label:'Desert Oasis Resort'}
 
-        
-    ]
 
     //https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find
     const obj = hotelData.find( (set)=>set.value ===props.review.hotel_Id)
