+++ front-end/src/data.js
@@ -1,0 +1,16 @@
+const hotelData= [
+    {value : 0 , label:'Grand Hotel'},
+    {value : 1 , label:'Ocean View Resort'},
+    {value : 2 , label:'Mountain Lodge'},
+    {value : 3 , label:'Lakeside Inn'},
+    {value : 4 , label:'Sunset Resort'},
+    {value : 5 , label:'Ski Chalet'},
+    {value : 6 , label:'Beachfront Paradise'},
+    {value : 7 , label:'City Center Hotel'},
+    {value : 8 , label:'Riverside Retreat'},
+    {value : 9 , label:'Desert Oasis Resort'}
+
+    
+]
+
+export default hotelData
